+++ src/main/java/com/gamesUP/gamesUP/repository/AvisRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 
-public interface AvisRepository extends JpaRepository<Avis,Integer> {
+public interface AvisRepository extends JpaRepository<Avis,Long> {
 
     List<Avis> findByGameId(Long gameId);
     List<Avis> findByUserId(Long userId);
