+++ src/main/java/com/gamesUP/gamesUP/repository/GameRepository.java
@@ -4,5 +4,5 @@ import com.gamesUP.gamesUP.model.Game;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-    public interface GameRepository extends JpaRepository<Game,Integer> {
+    public interface GameRepository extends JpaRepository<Game,Long> {
 }
