+++ src/main/java/com/gamesUP/gamesUP/repository/PurchaseLineRepository.java
@@ -3,5 +3,5 @@ package com.gamesUP.gamesUP.repository;
 import com.gamesUP.gamesUP.model.PurchaseLine;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PurchaseLineRepository extends CrudRepository<PurchaseLine,Integer> {
+public interface PurchaseLineRepository extends CrudRepository<PurchaseLine,Long> {
 }
