+++ src/main/java/com/gamesUP/gamesUP/repository/PurchaseRepository.java
@@ -4,5 +4,5 @@ import com.gamesUP.gamesUP.model.Purchase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface PurchaseRepository extends JpaRepository<Purchase,Integer> {
+public interface PurchaseRepository extends JpaRepository<Purchase,Long> {
 }
