+++ src/main/java/com/gamesUP/gamesUP/service/impl/AuthorServiceImpl.java
@@ -33,7 +33,7 @@ public class AuthorServiceImpl implements AuthorService {
     }
 
     @Override
-    public Author findById(int id_author) {
+    public Author findById(Long id_author) {
         Optional<Author> author = authorRepository.findById(id_author);
 
         //Si l'auteur est trouvé
@@ -45,18 +45,18 @@ public class AuthorServiceImpl implements AuthorService {
     }
 
     @Override
-    public void update(int id_author, Author author) {
+    public void update(Long id_author, Author author) {
 
-        author.setId((long) id_author);
+        author.setId( id_author);
         authorRepository.save(author);
     }
 
     @Override
     public void delete(Long id) {
-        if (!authorRepository.existsById(Math.toIntExact(id))) {
+        if (!authorRepository.existsById(id)) {
             throw new EntityDontExistException();
         }
-        authorRepository.deleteById(Math.toIntExact(id));
+        authorRepository.deleteById(id);
     }
